backend/pyproject.toml
[project]
name = "youtube-companion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.129.0",
    "google-api-python-client>=2.190.0",
    "google-auth>=2.48.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "google-genai>=1.63.0",
    "google-generativeai>=0.8.6",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.0",
    "sqlalchemy>=2.0.46",
]
